requirements.txt
anthropic==0.54.0
fastapi==0.115.12
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-core==0.3.65
langgraph==0.3.1
numpy==1.24.4
pydantic==2.11.5
python-dotenv==1.1.0
requests==2.32.3
uvicorn==0.34.3
structlog==23.1.0